+++ pcb-design/skidl/robot_dog_pcb/blocks/__init__.py
@@ -1,0 +1,12 @@
+from .expansion import ExpansionBlock
+from .mcu import McuBlock
+from .power import PowerBlock
+from .servo import ServoConnector, ServoConnectorBank
+
+__all__ = [
+    "ExpansionBlock",
+    "McuBlock",
+    "PowerBlock",
+    "ServoConnector",
+    "ServoConnectorBank",
+]